Fix compiler warning
git-svn-id: svn://scribus.net/trunk/Scribus@18559 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/pagepalette_widgets.cpp
@@ -498,7 +498,7 @@ SeItem* SeView::GetPageItem(int pageIndex)
 		{
 			QTableWidgetItem* tbItem = item(i, j);
 			SeItem* pageItem = dynamic_cast<SeItem*>(tbItem);
-			if (pageItem && pageItem->pageNumber == pageIndex)
+			if (pageItem && pageItem->pageNumber == static_cast<uint>(pageIndex))
 				return pageItem;
 		}
 	}
